backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
opencv-python==4.8.1.78
ultralytics
tensorflow==2.16.1
numpy==1.26.4
pillow
python-dotenv==1.0.0
psycopg2-binary==2.9.9
mediapipe==0.10.14